+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Unreleased
 
+## 1.10.0
+
 - Use Rust 1.92.0.
 - Update Ratatui to 0.30; fix error lifetimes.
 - Limit concurrency of git to max 5 concurrent operations.